Split Proxy
