+++ LeanplumSDK/LeanplumSDK/ClassesSwift/Notifications/NSObject+Notifications.swift
@@ -8,32 +8,43 @@
 import Foundation
 import UIKit
 
-// LP Swizzling looks for the selector in the same class
+/// LP Swizzling looks for the selector in the same class
 extension NSObject {
+
+    private var notificationsProxy: LeanplumNotificationsProxy {
+        return Leanplum.notificationsManager().proxy
+    }
+    
+    private var swizzling: LeanplumNotificationsProxy.Swizzled {
+        return notificationsProxy.swizzled
+    }
+    
     // MARK: - Register For Remote Notifications
-    @objc func leanplum_application(_ application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
+    @objc func leanplum_application(_ application: UIApplication,
+                                    didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
         Leanplum.notificationsManager().didRegisterForRemoteNotificationsWithDeviceToken(deviceToken)
         
         // Call overridden method
         let selector = #selector(self.leanplum_application(_:didRegisterForRemoteNotificationsWithDeviceToken:))
-        if Leanplum.notificationsManager().proxy.swizzledApplicationDidRegisterRemoteNotifications && self.responds(to: selector) {
+        if swizzling.applicationDidRegisterRemoteNotifications && self.responds(to: selector) {
             self.perform(selector, with: application, with: deviceToken)
         }
     }
     
-    @objc func leanplum_application(_ application: UIApplication, didFailToRegisterForRemoteNotificationsWithError error: Error) {
+    @objc func leanplum_application(_ application: UIApplication,
+                                    didFailToRegisterForRemoteNotificationsWithError error: Error) {
         Leanplum.notificationsManager().didFailToRegisterForRemoteNotificationsWithError(error)
         
         // Call overridden method
         let selector = #selector(self.leanplum_application(_:didFailToRegisterForRemoteNotificationsWithError:))
-        if Leanplum.notificationsManager().proxy.swizzledApplicationDidRegisterRemoteNotifications && self.responds(to: selector) {
+        if swizzling.applicationDidRegisterRemoteNotifications && self.responds(to: selector) {
             self.perform(selector, with: application, with: error)
         }
     }
     
     // MARK: - didReceiveRemoteNotification
     @objc func leanplum_application(_ application: UIApplication,
-                                     didReceiveRemoteNotification userInfo: [AnyHashable : Any],
+                                     didReceiveRemoteNotification userInfo: [AnyHashable: Any],
                                      fetchCompletionHandler completionHandler: @escaping (UIBackgroundFetchResult) -> Void) {
         Log.debug("Called swizzled application:didReceiveRemoteNotification:fetchCompletionHandler AppState: \(UIApplication.shared.applicationState.rawValue)")
         
@@ -44,7 +55,7 @@ extension NSObject {
                 var leanplumCompletionHandler:FetchResultCompletion?
                 let selector = #selector(self.leanplum_application(_:didReceiveRemoteNotification:fetchCompletionHandler:))
                 
-                if Leanplum.notificationsManager().proxy.swizzledApplicationDidReceiveRemoteNotificationWithCompletionHandler && self.responds(to: selector) {
+                if swizzling.applicationDidReceiveRemoteNotificationWithCompletionHandler && self.responds(to: selector) {
                     // Prevent calling the completionHandler twice
                     leanplumCompletionHandler = nil
                     // Call method directly since the it requires 3 objects and performSelector supports only 2
@@ -53,13 +64,13 @@ extension NSObject {
                 } else {
                     leanplumCompletionHandler = completionHandler
                     
-                    if Leanplum.notificationsManager().proxy.shouldFallbackToLegacyMethods {
-                        Leanplum.notificationsManager().proxy.appDelegate?.application?(UIApplication.shared, didReceiveRemoteNotification: userInfo)
+                    if notificationsProxy.shouldFallbackToLegacyMethods {
+                        notificationsProxy.appDelegate?.application?(UIApplication.shared, didReceiveRemoteNotification: userInfo)
                     }
                 }
                 
                 // Call completion handler
-                leanplumCompletionHandler?(Leanplum.notificationsManager().proxy.pushNotificationBackgroundFetchResult)
+                leanplumCompletionHandler?(notificationsProxy.pushNotificationBackgroundFetchResult)
             }
         }
         
@@ -74,7 +85,7 @@ extension NSObject {
             // Open notification will be handled by userNotificationCenter:didReceive or
             // application:didFinishLaunchingWithOptions
             // Receiving of notification when app is running on foreground is handled by userNotificationCenter:willPresent
-            if !Leanplum.notificationsManager().proxy.isEqualToHandledNotification(userInfo: userInfo) {
+            if !notificationsProxy.isEqualToHandledNotification(userInfo: userInfo) {
                 if state == .background {
                     Leanplum.notificationsManager().notificationReceived(userInfo: userInfo, isForeground: false)
                 }
@@ -91,7 +102,7 @@ extension NSObject {
         // iOS 9
         let state = UIApplication.shared.applicationState
         // Notification was not handled by application:didFinishLaunchingWithOptions
-        if !Leanplum.notificationsManager().proxy.isEqualToHandledNotification(userInfo: userInfo) {
+        if !notificationsProxy.isEqualToHandledNotification(userInfo: userInfo) {
         if  state == .inactive {
             // Open
             Leanplum.notificationsManager().notificationOpened(userInfo: userInfo)
@@ -99,7 +110,7 @@ extension NSObject {
             // There are cases where state has changed to active from inactive, when user tapped the notification
             // If app entered foreground right before calling this method, the app became active because notification was tapped
             // Otherwise, notification was received while app was active/foreground
-            if Leanplum.notificationsManager().proxy.resumedTimeInterval + 0.500 > NSDate().timeIntervalSince1970 {
+            if notificationsProxy.resumedTimeInterval + 0.500 > NSDate().timeIntervalSince1970 {
                 Leanplum.notificationsManager().notificationOpened(userInfo: userInfo)
             } else {
                 Leanplum.notificationsManager().notificationReceived(userInfo: userInfo, isForeground: true)
@@ -109,7 +120,7 @@ extension NSObject {
         }
         // App was waken up by notification, its receiving was handled by application:didFinishLaunchingWithOptions
         // didReceiveRemoteNotification is called again when user tapped it
-        } else if !Leanplum.notificationsManager().proxy.notificationOpenedFromStart && state != .background {
+        } else if !notificationsProxy.notificationOpenedFromStart && state != .background {
             Leanplum.notificationsManager().notificationOpened(userInfo: userInfo)
         }
     }
@@ -124,11 +135,11 @@ extension NSObject {
             // Call overridden method
             if LPUtils.isSwizzlingEnabled() {
                 let selector = #selector(self.leanplum_userNotificationCenter(_:didReceive:withCompletionHandler:))
-                if Leanplum.notificationsManager().proxy.swizzledUserNotificationCenterDidReceiveNotificationResponseWithCompletionHandler && self.responds(to: selector) {
+                if swizzling.userNotificationCenterDidReceiveNotificationResponseWithCompletionHandler && self.responds(to: selector) {
                     // Call original method
                     self.leanplum_userNotificationCenter(center, didReceive: response, withCompletionHandler: completionHandler)
                 } else {
-                    if Leanplum.notificationsManager().proxy.shouldFallbackToLegacyMethods {
+                    if notificationsProxy.shouldFallbackToLegacyMethods {
                         // Call iOS 10 UNUserNotificationCenter completionHandler from iOS 9's fetchCompletion handler
                         self.leanplum_application(UIApplication.shared, didReceiveRemoteNotification: userInfo, fetchCompletionHandler: { res in
                             completionHandler()
@@ -148,7 +159,7 @@ extension NSObject {
         
         // Handle UNNotificationDefaultActionIdentifier and Custom Actions
         if response.actionIdentifier != UNNotificationDismissActionIdentifier {
-            let notifWasOpenedFromStart = Leanplum.notificationsManager().proxy.isEqualToHandledNotification(userInfo: userInfo) && Leanplum.notificationsManager().proxy.notificationOpenedFromStart
+            let notifWasOpenedFromStart = notificationsProxy.isEqualToHandledNotification(userInfo: userInfo) && notificationsProxy.notificationOpenedFromStart
             if !notifWasOpenedFromStart {
                 // Open Notification action
                 if response.actionIdentifier == UNNotificationDefaultActionIdentifier {
@@ -162,16 +173,18 @@ extension NSObject {
     }
     
     @objc @available(iOS 10.0, *)
-    func leanplum_userNotificationCenter(_ center: UNUserNotificationCenter, willPresent notification: UNNotification, withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
+    func leanplum_userNotificationCenter(_ center: UNUserNotificationCenter,
+                                         willPresent notification: UNNotification,
+                                         withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
         Log.debug("Called swizzled userNotificationCenter:willPresentNotification:withCompletionHandler")
         
         defer {
             // Call overridden method
             let selector = #selector(self.leanplum_userNotificationCenter(_:willPresent:withCompletionHandler:))
-            if Leanplum.notificationsManager().proxy.swizzledUserNotificationCenterWillPresentNotificationWithCompletionHandler && self.responds(to: selector) {
+            if swizzling.userNotificationCenterWillPresentNotificationWithCompletionHandler && self.responds(to: selector) {
                 self.leanplum_userNotificationCenter(center, willPresent: notification, withCompletionHandler: completionHandler)
             } else {
-                completionHandler(Leanplum.notificationsManager().proxy.pushNotificationPresentationOption)
+                completionHandler(notificationsProxy.pushNotificationPresentationOption)
             }
         }
         
@@ -191,7 +204,7 @@ extension NSObject {
         defer {
             // Call overridden method
             let selector = #selector(self.leanplum_application(_:didReceive:))
-            if Leanplum.notificationsManager().proxy.swizzledApplicationDidReceiveLocalNotification && self.responds(to: selector) {
+            if swizzling.applicationDidReceiveLocalNotification && self.responds(to: selector) {
                 self.perform(selector, with: application, with: notification)
             }
         }
@@ -214,7 +227,7 @@ extension NSObject {
         
         // Call overridden method
         let selector = #selector(self.leanplum_application(_:didRegister:))
-        if Leanplum.notificationsManager().proxy.swizzledApplicationDidRegisterUserNotificationSettings && self.responds(to: selector) {
+        if swizzling.applicationDidRegisterUserNotificationSettings && self.responds(to: selector) {
             self.perform(selector, with: application, with: notificationSettings)
         }
     }
